+++ app/src/main/java/com/ably/tracking/demo/subscriber/Routes.kt
@@ -2,8 +2,8 @@ package com.ably.tracking.demo.subscriber
 
 sealed interface Routes {
 
-    object TrackableId : Routes {
-        const val pathWithParams = "trackableId/"
+    object CreateOrder : Routes {
+        const val pathWithParams = "order/create/"
     }
 
     object Dashboard : Routes {
